modifications in the interface for brnns
