+++ helix/components/forms/__init__.py
@@ -1,0 +1,11 @@
+"""Forms package for Helix components.
+
+This package contains form components for configuring various aspects of Helix:
+- Machine Learning model options
+- Data preprocessing options
+- Feature importance options
+"""
+
+from helix.components.forms.forms_ml_opts import ml_options_form
+
+__all__ = ["ml_options_form"]
